Refactored into code - removed unnecessary imports; renamed files;
