+++ frontend/src/components/SharedPopover.tsx
@@ -1,6 +1,6 @@
 import Popover, { PopoverPlacement } from "react-native-popover-view";
 import { StyleSheet, TouchableHighlight, TouchableOpacity, View, Text } from "react-native";
-import { SharedPopoverProps } from "../types/Popover";
+import { SharedPopoverProps } from "../types/SharedPopover";
 import { RefObject } from "react";
 
 const SharedPopover = ({ word, anchorRef, isVisible, onClose }: SharedPopoverProps) => {
