+++ frontend/src/types/Word.ts
@@ -1,8 +1,14 @@
-import wordInfo from "../mock_data/word_data";
+import { RefObject } from "react";
+import { View } from "react-native";
 
 export interface WordInfo {
     word: string
     definition: string;
     type: string;
     example?: string;
+}
+
+export interface ClickableWordProps {
+    word: WordInfo;
+    onWordPress: (word: WordInfo, ref: RefObject<View>) => void;
 }
